Fix for Grenadier and some other related stuff
Changes spalsh damage from 6 into 128.
NOTE: 6 is the value taken as it is from the original game.

Changes inaccuracy from 813 into 600 as it was too inaccurate vs the original game.

Also Changes LaunchAngle into 92 from 62 as a visual change.

NOTE: This is not fully tested with the original game, further tests are adviced.

Closes #87